[REFACTOR] #199: in,exterior 공통 컴포넌트 분리
