+++ frontend/src/containers/ExteriorPage/ExteriorSelectContainer.tsx
@@ -1,10 +1,9 @@
-import { styled, useTheme } from 'styled-components';
-import { BodyKrRegular3, HeadingKrMedium5 } from '../../styles/typefaces';
+import { ReactNode, useState } from 'react';
+import { styled } from 'styled-components';
 import CenterWrapper from '../../components/layout/CenterWrapper';
-import { useState } from 'react';
-import { ArrowLeft, ArrowRight } from '../../components/common/icons/Icons';
 import PriceSummary from '../../components/summary/PriceSummary';
 import ExteriorCard from '../../components/cards/ExteriorCard';
+import CardSlider from '../../components/cardSlider/CardSlider';
 import { MAX_PAGE, NUM_IN_A_PAGE } from '../../utils/constants';
 
 interface ISelected {
@@ -13,26 +12,9 @@ interface ISelected {
 }
 
 export default function ExteriorSelectContainer() {
-  const [page, setPage] = useState(0);
   const [selectedIdx, setSelectedIdx] = useState<ISelected>({ page: 0, idx: 0 });
-  const theme = useTheme();
-
   const cardIndices = Array.from({ length: NUM_IN_A_PAGE }, (_, index) => index + 1);
-  const arrowLeftColor = page <= 0 ? theme.color.gray200 : theme.color.gray600;
-  const arrowRightColor = page >= MAX_PAGE - 1 ? theme.color.gray200 : theme.color.gray600;
 
-  const handlePageNext = () => {
-    setPage((cur) => {
-      const nextPage = cur + 1 >= MAX_PAGE ? cur : cur + 1;
-      return nextPage;
-    });
-  };
-  const handlePagePrev = () => {
-    setPage((cur) => {
-      const prevPage = cur - 1 < 0 ? cur : cur - 1;
-      return prevPage;
-    });
-  };
   const handleSelectedIdx = ({ page, idx }: ISelected) => {
     setSelectedIdx({ page, idx });
   };
@@ -40,7 +22,7 @@ export default function ExteriorSelectContainer() {
     return page === selectedIdx.page && idx === selectedIdx.idx;
   };
 
-  const CardPageList = [];
+  const CardPageList: ReactNode[] = [];
   for (let i = 0; i < MAX_PAGE; i++) {
     const newCardPage = (
       <CardPage key={i}>
@@ -62,21 +44,7 @@ export default function ExteriorSelectContainer() {
 
   return (
     <Wrapper>
-      <Header>
-        <Title>외장 색상을 선택 해주세요</Title>
-        <PageButtonWrapper>
-          <PageButton onClick={handlePagePrev}>
-            <ArrowLeft fill={arrowLeftColor} />
-          </PageButton>
-          <Page>
-            {page + 1}/{MAX_PAGE}
-          </Page>
-          <PageButton onClick={handlePageNext}>
-            <ArrowRight fill={arrowRightColor} />
-          </PageButton>
-        </PageButtonWrapper>
-      </Header>
-      <SelectSection>{CardPageList[page]}</SelectSection>
+      <CardSlider title="외장 색상을 선택해주세요." cardList={CardPageList} />
       <Footer>
         <PriceSummary />
       </Footer>
@@ -87,29 +55,6 @@ export default function ExteriorSelectContainer() {
 const Wrapper = styled(CenterWrapper)`
   display: flex;
   flex-direction: column;
-`;
-const Header = styled.div`
-  margin-top: 16px;
-
-  display: flex;
-  justify-content: space-between;
-`;
-const Title = styled.div`
-  ${HeadingKrMedium5};
-`;
-const PageButtonWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 116px;
-  ${BodyKrRegular3}
-  color: ${({ theme }) => theme.color.gray500};
-`;
-const PageButton = styled.button``;
-const Page = styled.span``;
-
-const SelectSection = styled.div`
-  position: relative;
 `;
 const CardPage = styled.div`
   display: flex;
